+++ nodes/PineconeAssistant/v1/actions/deleteFile.test.ts
