+++ nodes/PineconeAssistant/v1/actions/deleteFile.ts
