+++ nodes/PineconeAssistant/v1/actions/listFiles.test.ts
