+++ nodes/PineconeAssistant/v1/actions/listFiles.ts
