+++ nodes/PineconeAssistant/v1/actions/uploadFile.ts
