+++ nodes/PineconeAssistant/v1/genericFunctions.ts
